requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
tortoise-orm[asyncpg]==0.21.6
aerich==0.7.2
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
python-dotenv==1.0.0
aiofiles==24.1.0
numpy==1.26.4
scipy==1.11.4
python-docx==1.1.0
openpyxl==3.1.2
matplotlib==3.8.2
